naming and typing
