+++ test/unit/DocumentsBatch.spec.ts
@@ -3,8 +3,9 @@ import {
   DocumentCreateTransitionWASM,
   DocumentDeleteTransitionWASM,
   DocumentPurchaseTransitionWASM, DocumentReplaceTransitionWASM,
-  DocumentsBatchWASM, DocumentTransferTransitionWASM, DocumentUpdatePriceTransitionWASM
+  DocumentsBatchWASM, DocumentTransferTransitionWASM, DocumentUpdatePriceTransitionWASM, StateTransitionWASM
 } from 'pshenmic-dpp'
+import {create} from "dashhd";
 
 let sdk: DashPlatformSDK
 
@@ -35,18 +36,15 @@ describe('DocumentsBatch', () => {
   })
 
   test('should be able to create document batch from document create transition', async () => {
+
     const document = await sdk.documents.create(dataContract, documentType, data, identityContractNonce, identity)
 
     const transition = new sdk.wasm.DocumentCreateTransitionWASM(document, identityContractNonce, document.getDocumentTypeName())
 
     const batch = await sdk.stateTransitions.documentsBatch.create(transition, identity)
 
-    const batchFromArray = await sdk.stateTransitions.documentsBatch.create([transition, transition], identity)
-
     expect(transition).toEqual(expect.any(DocumentCreateTransitionWASM))
-    expect(batch).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batchFromArray).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batchFromArray.transitions.length).toEqual(2)
+    expect(batch).toEqual(expect.any(StateTransitionWASM))
   })
 
   test('should be able to create document batch from document delete transition', async () => {
@@ -56,12 +54,8 @@ describe('DocumentsBatch', () => {
 
     const batch = await sdk.stateTransitions.documentsBatch.create(transition, identity)
 
-    const batchFromArray = await sdk.stateTransitions.documentsBatch.create([transition, transition], identity)
-
     expect(transition).toEqual(expect.any(DocumentDeleteTransitionWASM))
-    expect(batch).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batchFromArray).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batchFromArray.transitions.length).toEqual(2)
+    expect(batch).toEqual(expect.any(StateTransitionWASM))
   })
 
   test('should be able to create document batch from document purchase transition', async () => {
@@ -69,14 +63,10 @@ describe('DocumentsBatch', () => {
 
     const transition = new sdk.wasm.DocumentPurchaseTransitionWASM(document, identityContractNonce, document.getDocumentTypeName(), price)
 
-    const batch = await sdk.stateTransitions.documentsBatch.create(transition, identity)
-
-    const batchFromArray = await sdk.stateTransitions.documentsBatch.create([transition, transition], identity)
+    const batch = await sdk.stateTransitions.documentsBatch.create(document, identity)
 
     expect(transition).toEqual(expect.any(DocumentPurchaseTransitionWASM))
-    expect(batch).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batchFromArray).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batchFromArray.transitions.length).toEqual(2)
+    expect(batch).toEqual(expect.any(StateTransitionWASM))
   })
 
   test('should be able to create document batch from document replace transition', async () => {
@@ -86,12 +76,8 @@ describe('DocumentsBatch', () => {
 
     const batch = await sdk.stateTransitions.documentsBatch.create(transition, identity)
 
-    const batchFromArray = await sdk.stateTransitions.documentsBatch.create([transition, transition], identity)
-
     expect(transition).toEqual(expect.any(DocumentReplaceTransitionWASM))
-    expect(batch).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batchFromArray).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batchFromArray.transitions.length).toEqual(2)
+    expect(batch).toEqual(expect.any(StateTransitionWASM))
   })
 
   test('should be able to create document batch from document transfer transition', async () => {
@@ -101,12 +87,8 @@ describe('DocumentsBatch', () => {
 
     const batch = await sdk.stateTransitions.documentsBatch.create(transition, identity)
 
-    const batchFromArray = await sdk.stateTransitions.documentsBatch.create([transition, transition], identity)
-
     expect(transition).toEqual(expect.any(DocumentTransferTransitionWASM))
-    expect(batch).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batchFromArray).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batchFromArray.transitions.length).toEqual(2)
+    expect(batch).toEqual(expect.any(StateTransitionWASM))
   })
 
   test('should be able to create document batch from document update price transition', async () => {
@@ -116,12 +98,8 @@ describe('DocumentsBatch', () => {
 
     const batch = await sdk.stateTransitions.documentsBatch.create(transition, identity)
 
-    const batchFromArray = await sdk.stateTransitions.documentsBatch.create([transition, transition], identity)
-
     expect(transition).toEqual(expect.any(DocumentUpdatePriceTransitionWASM))
-    expect(batch).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batchFromArray).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batchFromArray.transitions.length).toEqual(2)
+    expect(batch).toEqual(expect.any(StateTransitionWASM))
   })
 
   test('should be able to create document batch from document', async () => {
@@ -129,26 +107,6 @@ describe('DocumentsBatch', () => {
 
     const batch = await sdk.stateTransitions.documentsBatch.create(document, identity, { identityContractNonce: BigInt(1) })
 
-    const batchFromArray = await sdk.stateTransitions.documentsBatch.create([document, document], identity, { identityContractNonce: BigInt(1) })
-
-    expect(batch).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batchFromArray).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batchFromArray.transitions.length).toEqual(2)
-  })
-
-  test('should be able to create document batch from mixed input', async () => {
-    const document = await sdk.documents.create(dataContract, documentType, data, identityContractNonce, identity)
-
-    const transitionCreate = new sdk.wasm.DocumentCreateTransitionWASM(document, identityContractNonce, document.getDocumentTypeName())
-    const transitionDelete = new sdk.wasm.DocumentDeleteTransitionWASM(document, identityContractNonce, document.getDocumentTypeName())
-    const transitionPurchase = new sdk.wasm.DocumentPurchaseTransitionWASM(document, identityContractNonce, document.getDocumentTypeName(), price)
-    const transitionReplace = new sdk.wasm.DocumentReplaceTransitionWASM(document, identityContractNonce, document.getDocumentTypeName())
-    const transitionTransfer = new sdk.wasm.DocumentTransferTransitionWASM(document, identityContractNonce, document.getDocumentTypeName(), recipient)
-    const transitionUpdatePrice = new sdk.wasm.DocumentUpdatePriceTransitionWASM(document, identityContractNonce, document.getDocumentTypeName(), price)
-
-    const batch = await sdk.stateTransitions.documentsBatch.create([document, transitionCreate, transitionDelete, transitionPurchase, transitionReplace, transitionTransfer, transitionUpdatePrice], identity, { identityContractNonce: BigInt(1) })
-
-    expect(batch).toEqual(expect.any(DocumentsBatchWASM))
-    expect(batch.transitions.length).toEqual(7)
+    expect(batch).toEqual(expect.any(StateTransitionWASM))
   })
 })
